requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai
textblob
spacy
pydantic
pydantic-settings
pandas
PyYAML